both calls and method annotations for class level
