+++ src/main/java/io/lenar/easy/log/EasyLogger.java
@@ -5,6 +5,8 @@ import io.lenar.easy.log.annotations.LogCall;
 import io.lenar.easy.log.annotations.LogCalls;
 import io.lenar.easy.log.annotations.LogMethod;
 import io.lenar.easy.log.annotations.LogMethods;
+import io.lenar.easy.log.support.JoinPointLogger;
+
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
@@ -18,22 +20,22 @@ public class EasyLogger extends JoinPointLogger {
 
     @Around("anyMethod() && @annotation(annotation)")
     public Object logCall(ProceedingJoinPoint jp, LogCall annotation) throws Throwable {
-        return logCallMethodLevel(jp, annotation);
+        return logAsCall(jp, annotation.name());
     }
 
     @Around("anyMethod() && @within(annotation)")
     public Object logCalls(ProceedingJoinPoint jp, LogCalls annotation) throws Throwable {
-        return logCallClassLevel(jp, annotation);
+        return logAsCall(jp, annotation.name());
     }
 
     @Around("anyMethod() && @annotation(annotation)")
     public Object logMethod(ProceedingJoinPoint jp, LogMethod annotation) throws Throwable {
-        return logMethodMethodLevel(jp, annotation);
+        return logAsMethod(jp);
     }
 
     @Around("anyMethod() && @within(annotation)")
     public Object logMethods(ProceedingJoinPoint jp, LogMethods annotation) throws Throwable {
-        return logMethodClassLevel(jp, annotation);
+        return logAsMethod(jp);
     }
